3DSOLID: more corrections to member types of dxf_3dsolid_struct.

+++ src/3dsolid.h
@@ -161,7 +161,7 @@ dxf_3dsolid_struct
                 /*!< Hard pointer ID / handle of PlotStyleName object.\n
                  * Group code = 390.\n
                  * \since Introduced in version R2009. */
-        long color_value;
+        int32_t color_value;
                 /*!< A 24-bit color value that should be dealt with in
                  * terms of bytes with values of 0 to 255.\n
                  * The lowest byte is the blue value, the middle byte is
@@ -181,7 +181,7 @@ dxf_3dsolid_struct
                  * class-level transparency data.\n
                  * Group code = 430.\n
                  * \since Introduced in version R2004. */
-        long transparency;
+        int32_t transparency;
                 /*!< Transparency value.\n
                  * The group code cannot be used by custom entities for
                  * their own data because the group code is reserved for
@@ -254,12 +254,12 @@ int16_t dxf_3dsolid_get_lineweight (Dxf3dsolid *solid);
 Dxf3dsolid *dxf_3dsolid_set_lineweight (Dxf3dsolid *solid, int16_t lineweight);
 char *dxf_3dsolid_get_plot_style_name (Dxf3dsolid *solid);
 Dxf3dsolid *dxf_3dsolid_set_plot_style_name (Dxf3dsolid *solid, char *plot_style_name);
-long dxf_3dsolid_get_color_value (Dxf3dsolid *solid);
-Dxf3dsolid *dxf_3dsolid_set_color_value (Dxf3dsolid *solid, long color_value);
+int32_t dxf_3dsolid_get_color_value (Dxf3dsolid *solid);
+Dxf3dsolid *dxf_3dsolid_set_color_value (Dxf3dsolid *solid, int32_t color_value);
 char *dxf_3dsolid_get_color_name (Dxf3dsolid *solid);
 Dxf3dsolid *dxf_3dsolid_set_color_name (Dxf3dsolid *solid, char *color_name);
-long dxf_3dsolid_get_transparency (Dxf3dsolid *solid);
-Dxf3dsolid *dxf_3dsolid_set_transparency (Dxf3dsolid *solid, long transparency);
+int32_t dxf_3dsolid_get_transparency (Dxf3dsolid *solid);
+Dxf3dsolid *dxf_3dsolid_set_transparency (Dxf3dsolid *solid, int32_t transparency);
 DxfProprietaryData *dxf_3dsolid_get_proprietary_data (Dxf3dsolid *solid);
 Dxf3dsolid *dxf_3dsolid_set_proprietary_data (Dxf3dsolid *solid, DxfProprietaryData *proprietary_data);
 DxfProprietaryData *dxf_3dsolid_get_additional_proprietary_data (Dxf3dsolid *solid);
